requirements.txt
python-dotenv
uvicorn
gunicorn
mysql-connector-python
fastapi
sentry-sdk
requests
bs4
APScheduler
matplotlib
pre-commit
flake8
black
pydocstyle
mypy
isort